feat: サービスの統合とリファクタリングを実施し、不要なコードを削除
