+++ src/services/history/index.ts
@@ -1,4 +1,4 @@
 export { HistoryService } from './service';
-export * as HistoryTypes from './history.types';
+export * from './types';
 export * from './converter';
-export * as HistoryUtils from './history.utils';
+export * as HistoryUtils from './utils';
